finished porting chatStream tests
Had to make the util function handleStreamResponse a class method so it could be mocked
