+++ tests/dep.js
@@ -15,77 +15,4 @@ describe("Mistral Client", () => {
   beforeEach(() => {
     client = new MistralClient();
   });
-
-  describe("chatStream()", () => {
-    it("should return parsed, streamed response", async () => {
-      // Mock the fetch function
-      const mockResponse = mockChatResponseStreamingPayload();
-      globalThis.fetch = mockFetchStream(200, mockResponse);
-
-      const response = await client.chatStream({
-        model: "mistral-small",
-        messages: [
-          {
-            role: "user",
-            content: "What is the best French cheese?",
-          },
-        ],
-      });
-
-      const parsedResponse = [];
-      for await (const r of response) {
-        parsedResponse.push(r);
-      }
-
-      expect(parsedResponse.length).toEqual(11);
-    });
-
-    it("should return parsed, streamed response with safeMode", async () => {
-      // Mock the fetch function
-      const mockResponse = mockChatResponseStreamingPayload();
-      globalThis.fetch = mockFetchStream(200, mockResponse);
-
-      const response = await client.chatStream({
-        model: "mistral-small",
-        messages: [
-          {
-            role: "user",
-            content: "What is the best French cheese?",
-          },
-        ],
-        safeMode: true,
-      });
-
-      const parsedResponse = [];
-      for await (const r of response) {
-        parsedResponse.push(r);
-      }
-
-      expect(parsedResponse.length).toEqual(11);
-    });
-
-    it("should return parsed, streamed response with safePrompt", async () => {
-      // Mock the fetch function
-      const mockResponse = mockChatResponseStreamingPayload();
-      globalThis.fetch = mockFetchStream(200, mockResponse);
-
-      const response = await client.chatStream({
-        model: "mistral-small",
-        messages: [
-          {
-            role: "user",
-            content: "What is the best French cheese?",
-          },
-        ],
-        safePrompt: true,
-      });
-
-      const parsedResponse = [];
-      for await (const r of response) {
-        parsedResponse.push(r);
-      }
-
-      expect(parsedResponse.length).toEqual(11);
-    });
-  });
 });
